requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.0
backcall==0.2.0
click==8.1.7
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.104.0
h11==0.14.0
idna==3.4
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.4.0
jupyter_core==5.4.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
wcwidth==0.2.8
